move FRDFileHeader definition into a separate header
